# gateway/openapi.yaml
openapi: 3.0.3
info:
  title: RudderStack HTTP API
  version: 1.0.0-oas3
  description: >-
    The RudderStack HTTP API lets you track your customer event data and route
    it to your desired destinations.
  termsOfService: https://www.rudderstack.com/master-service-agreement/
  contact:
    email: support@rudderstack.com
  license:
    name: Elastic License 2.0
    url: https://www.elastic.co/licensing/elastic-license
paths:
  /v1/identify:
    post:
      tags:
        - HTTP API
      summary: Identify
      description: >-
        The identify call lets you associate a visiting user to their actions
        and record any associated traits.
      operationId: Identify
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentifyPayload'
        required: true
      responses:
        '200':
          description: StatusOK
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "OK"
        '400':
          description: StatusBadRequest
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Invalid request"
        '401':
          description: StatusUnauthorized
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Invalid Authorization Header"
        '404':
          description: StatusNotFound
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Source does not accept webhook events"
        '413':
          description: StatusRequestEntityTooLarge
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Request size too large"
        '429':
          description: StatusTooManyRequests
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Too many requests"
      security:
        - writeKeyAuth: []
...
  /internal/v1/extract:
    post:
      tags:
        - Internal API
      summary: Extract
      description: Handles extract requests.
      operationId: Extract
      responses:
        '200':
          description: StatusOK
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "OK"
        '400':
          description: StatusBadRequest
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Invalid request"
        '401':
          description: StatusUnauthorized
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Invalid Authorization Header"
        '404':
          description: StatusNotFound
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Source does not accept webhook events"
        '413':
          description: StatusRequestEntityTooLarge
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Request size too large"
        '429':
          description: StatusTooManyRequests
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Too many requests"
      security:
        - writeKeyAuth: []
...
  /internal/v1/audiencelist:
    post:
      tags:
        - Internal API
      summary: Audience List
      description: Handles audience list requests.
      operationId: AudienceList
      responses:
        '200':
          description: StatusOK
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "OK"
        '400':
          description: StatusBadRequest
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Invalid request"
        '401':
          description: StatusUnauthorized
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Invalid Authorization Header"
        '404':
          description: StatusNotFound
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Source does not accept webhook events"
        '413':
          description: StatusRequestEntityTooLarge
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Request size too large"
        '429':
          description: StatusTooManyRequests
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Too many requests"
      security:
        - writeKeyAuth: []
...
  /internal/v1/batch:
    post:
      tags:
        - Internal API
      summary: Batch
      description: Handles internal batch requests.
      operationId: InternalBatch
      responses:
        '200':
          description: StatusOK
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "OK"
        '400':
          description: StatusBadRequest
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Invalid request"
        '401':
          description: StatusUnauthorized
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
              example: "Invalid Authorization Header"
      security:
        - sourceIDAuth: []
  # ============================================================
  # Functions API Endpoints (Sprint 4-6, E-015, E-018)
  # ============================================================
...
servers:
  - url: /v1
components:
  securitySchemes:
    writeKeyAuth:
      type: http
      scheme: basic
      description: Write Key Basic Authentication
    sourceIDAuth:
      type: http
      scheme: basic
      description: SourceID Basic Authentication
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Bearer token authentication for Functions, Protocols, and Profiles
        management APIs.
    functionsTokenAuth:
      type: apiKey
      in: header
      name: X-Functions-Token
      description: >-
        Function-specific authentication token for Source Functions webhook
        ingestion endpoints.
  schemas:
    UserAgentData:
      type: object
      description: >-
        Structured Client Hints API data from the browser. Contains
        low-entropy fields (brands, mobile, platform) and optional
        high-entropy fields.
      properties:
        brands:
          type: array
          description: Array of brand/version objects from the User-Agent Client Hints API.
          items:
            type: object
            properties:
              brand:
                type: string
                description: Brand name.
              version:
                type: string
                description: Brand version string.
        mobile:
          type: boolean
          description: Whether the device is a mobile device.
        platform:
          type: string
          description: Operating system platform name.
        bitness:
          type: string
          description: CPU architecture bitness (e.g., "64").
        model:
          type: string
          description: Device model identifier.
        platformVersion:
          type: string
          description: Operating system version.
        uaFullVersion:
          type: string
          description: Full browser version string.
        fullVersionList:
          type: array
          description: Array of brand/version objects with full version information.
          items:
            type: object
            properties:
              brand:
                type: string
                description: Brand name.
              version:
                type: string
                description: Full version string.
        wow64:
          type: boolean
          description: Whether the browser is running in WoW64 mode.
    IdentifyPayload:
      type: object
      properties:
        userId:
          type: string
          description: Unique identifier for a particular user in your database.
        anonymousId:
          type: string
          description: >-
            Sets the user ID for cases where there is no unique identifier for
            the user.
        context:
          type: object
          description: Dictionary of information that provides context about a message.
          properties:
            traits:
              type: object
              properties:
                trait1:
                  type: string
                  description: New value for trait1.
            ip:
              type: string
              description: IP address.
            library:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the library.
            userAgent:
              type: string
              description: User agent string of the device making the request.
            userAgentData:
              $ref: '#/components/schemas/UserAgentData'
            channel:
              type: string
              description: >-
                Channel of the event source. Auto-populated by SDKs as
                "server", "browser", or "mobile".
              enum:
                - server
                - browser
                - mobile
        timestamp:
          type: string
          format: date-time
          description: The timestamp of the message’s arrival.
...